G9-69 feat: add list group
