+++ package.json
@@ -17,7 +17,7 @@
 		"@emotion/react": "^11.11.1",
 		"@emotion/styled": "^11.11.0",
 		"@mui/icons-material": "^5.14.16",
-		"@mui/material": "^5.14.16",
+		"@mui/material": "^5.14.17",
 		"@reduxjs/toolkit": "^1.9.7",
 		"js-cookie": "^3.0.5",
 		"react": "^18.2.0",
