+++ src/features/groups/api.ts
@@ -1,0 +1,6 @@
+import Group from './types/Group';
+
+export async function getAll(): Promise<Group[]> {
+	const result = await fetch(`/api/groups/`);
+	return result.json();
+}
