+++ src/features/groups/types/Group.ts
@@ -1,0 +1,6 @@
+export default interface Group {
+	id: number;
+	name: string;
+	courseId: number;
+	archived: boolean;
+}
