+++ src/features/groups/types/GroupsState.ts
@@ -1,0 +1,6 @@
+import Group from './Group';
+
+export default interface GroupsState {
+	groups: Group[];
+	error?: string;
+}
